* REFACTOR: Moved method. * ADD: Added empty newsletter email.
