+++ backend/src/main/java/com/huellapositiva/application/controller/VolunteerApiController.java
@@ -62,6 +62,8 @@ public class VolunteerApiController {
 
     private final UploadPhotoAction uploadPhotoAction;
 
+    private final ChangeStatusNewsletterSubscriptionAction changeStatusNewsletterSubscriptionAction;
+
     @Operation(
             summary = "Register a new volunteer",
             description = "Register a new volunteer",
@@ -290,5 +292,38 @@ public class VolunteerApiController {
     public void updateProfileInformation(@Validated @RequestBody UpdateProfileRequestDto updateProfileRequestDto,
                                          @Parameter(hidden = true) @AuthenticationPrincipal String volunteerEmail) {
         updateVolunteerProfileAction.execute(updateProfileRequestDto, volunteerEmail);
+    }
+
+    @Operation(
+            summary = "Change status of subscription to newsletter",
+            description = "Changes the status of the subscribed parameter on the specified volunteer",
+            tags = "newsletter",
+            parameters = {
+                    @Parameter(name = "X-XSRF-TOKEN", in = ParameterIn.HEADER, required = true, example = "a6f5086d-af6b-464f-988b-7a604e46062b", description = "For take this value, open your inspector code on your browser, and take the value of the cookie with the name 'XSRF-TOKEN'. Example: a6f5086d-af6b-464f-988b-7a604e46062b"),
+                    @Parameter(name = "XSRF-TOKEN", in = ParameterIn.COOKIE, required = true, example = "a6f5086d-af6b-464f-988b-7a604e46062b", description = "Same value of X-XSRF-TOKEN")
+            },
+            security = {
+                    @SecurityRequirement(name = "accessToken")
+            }
+    )
+    @ApiResponses(
+            value = {
+                    @ApiResponse(
+                            responseCode = "200",
+                            description = "Ok, status of subscribed field changed successfully"
+                    ),
+                    @ApiResponse(
+                            responseCode = "500",
+                            description = "Internal server error, could not fetch the user data due to a connectivity issue."
+                    )
+            }
+    )
+    @PostMapping("/profile/newsletter")
+    @RolesAllowed("VOLUNTEER")
+    @ResponseBody
+    @ResponseStatus(HttpStatus.OK)
+    public void changeStatusNewsletterSubscription(@RequestBody Boolean subscribed,
+                                                   @Parameter(hidden = true)@AuthenticationPrincipal String volunteerEmail) {
+        changeStatusNewsletterSubscriptionAction.execute(subscribed, volunteerEmail);
     }
 }
