+++ backend/src/main/java/com/huellapositiva/domain/actions/ManageNewsletterExcelAction.java
@@ -33,7 +33,7 @@ public class ManageNewsletterExcelAction {
     public void execute(String email) throws IOException {
         List<JpaVolunteer> subscribedVolunteers = jpaVolunteerRepository.findSubscribedVolunteers();
         if (subscribedVolunteers.isEmpty()){
-            throw new IllegalStateException("Can not find any volunteers subscribed to newsletter");
+            communicationService.sendEmptyNewsletter(EmailAddress.from(email));
         } else {
             buildExcel(subscribedVolunteers);
             URL url = uploadExcelAndGetUrl();
