+++ backend/src/main/java/com/huellapositiva/infrastructure/TemplateService.java
@@ -104,4 +104,10 @@ public class TemplateService {
         variables.put("NEWSLETTER_URL", stringUrl);
         return new EmailTemplate(template).parse(variables);
     }
+
+    public EmailTemplate getEmptyNewsletterEmailTemplate() {
+        String relativePath = "classpath:templates/emails/emptyNewsletter.txt";
+        String template = getFileContent(relativePath);
+        return new EmailTemplate(template);
+    }
 }
