+++ backend/src/main/resources/templates/emails/emptyNewsletter.txt
@@ -1,0 +1,3 @@
+No se han encontrado voluntarios suscritos a la newsletter, por lo que no ha sido posible generar el documento excel.
+
+Un cordial saludo.
