+++ backend/src/test/java/com/huellapositiva/api/NewsletterControllerShould.java
@@ -1,10 +1,7 @@
 package com.huellapositiva.api;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.huellapositiva.application.dto.JwtResponseDto;
 import com.huellapositiva.domain.model.valueobjects.Roles;
-import com.huellapositiva.infrastructure.orm.entities.JpaVolunteer;
-import com.huellapositiva.infrastructure.orm.repository.JpaVolunteerRepository;
 import com.huellapositiva.util.TestData;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -20,10 +17,7 @@ import java.util.UUID;
 
 import static com.huellapositiva.util.TestData.*;
 import static com.huellapositiva.util.TestUtils.loginAndGetJwtTokens;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -31,7 +25,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @Import(TestData.class)
 class NewsletterControllerShould {
     private static final String NEWSLETTER_URL = "/api/v1/newsletter";
-    private static final ObjectMapper objectMapper = new ObjectMapper();
 
     @Autowired
     private TestData testData;
@@ -39,39 +32,9 @@ class NewsletterControllerShould {
     @Autowired
     private MockMvc mvc;
 
-    @Autowired
-    private JpaVolunteerRepository jpaVolunteerRepository;
-
     @BeforeEach
     void beforeEach() {
         testData.resetData();
-    }
-
-    @Test
-    void return_200_when_state_of_subscribed_field_changed_successfully() throws Exception {
-        testData.createVolunteerWithProfile(DEFAULT_EMAIL, DEFAULT_PASSWORD);
-
-        JwtResponseDto jwtResponseDto = loginAndGetJwtTokens(mvc, DEFAULT_EMAIL, DEFAULT_PASSWORD);
-
-        mvc.perform(post(NEWSLETTER_URL + "/changeStatusNewsletterSubscription")
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + jwtResponseDto.getAccessToken())
-                .content(objectMapper.writeValueAsString(Boolean.TRUE))
-                .contentType(MediaType.APPLICATION_JSON)
-                .with(csrf())
-                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk());
-        JpaVolunteer volunteer = jpaVolunteerRepository.findByEmail(DEFAULT_EMAIL).orElseThrow();
-        assertThat(volunteer.getProfile().isSubscribed()).isTrue();
-
-        mvc.perform(post(NEWSLETTER_URL + "/changeStatusNewsletterSubscription")
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + jwtResponseDto.getAccessToken())
-                .content(objectMapper.writeValueAsString(Boolean.FALSE))
-                .contentType(MediaType.APPLICATION_JSON)
-                .with(csrf())
-                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk());
-        volunteer = jpaVolunteerRepository.findByEmail(DEFAULT_EMAIL).orElseThrow();
-        assertThat(volunteer.getProfile().isSubscribed()).isFalse();
     }
 
     @Test
@@ -88,19 +51,6 @@ class NewsletterControllerShould {
                 .contentType(MediaType.APPLICATION_JSON)
                 .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isNoContent())
-                .andReturn().getResponse();
-    }
-
-    @Test
-    void return_404_when_not_found_any_volunteers_subscribed_to_newsletter() throws Exception {
-        testData.createCredential("revisor@huellapositiva.com", UUID.randomUUID(), DEFAULT_PASSWORD, Roles.REVISER);
-        JwtResponseDto jwtResponseDto = loginAndGetJwtTokens(mvc, "revisor@huellapositiva.com", DEFAULT_PASSWORD);
-
-        mvc.perform(get(NEWSLETTER_URL + "/getNewsletterExcel")
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + jwtResponseDto.getAccessToken())
-                .contentType(MediaType.APPLICATION_JSON)
-                .accept(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNotFound())
                 .andReturn().getResponse();
     }
 }
